flatness_detection/requirements.txt
Flask
Flask-Cors
torch
torchvision
numpy==1.23.4
opencv-python
Pillow
requests
xlwt
scikit-image
pymysql
cryptography
matplotlib
